HZN-1212: Move rest impelmentation to osgi container
